update youtube playlist ui
There are still some issues with using the site on small or medium screens. Not sure how i want to fix that issue.
I might just have to reorganize things, maybe detach the video from the right side so it can be set higher up or something.
We could also make it a select so we either show the playlists or the videos in the playlists.